SFTP Plugin: refactor for better stability
Refactor to watch the global GVolumeMonitor instead of the result of
GMount operations and the eventual GMount::unmounted signals. Also avoid
unmounting implicity in most cases.

This takes advantage of a number of quirks that occur with the Android
SFTP server, allowing mounts to survive device connection breaks and be
reused in several cases.

closes #478